+++ R/testthat-ui.R
@@ -67,6 +67,11 @@
 #' manually assess whether changes reflect actual problems in your
 #' package.
 #'
+#' If you need to override the default vdiffr behaviour on CRAN (not
+#' recommended) or Travis (for example to run the tests in a
+#' particular builds but not others), set the `VDIFFR_RUN_TESTS`
+#' environment variable to "true" or "false".
+#'
 #' @section Debugging:
 #'
 #' It is sometimes difficult to understand the cause of a failure.